Transform main.rs into example
Previously, the crate included a main.rs file that parsed two eml failes
and printed some headers to the standard output.  As this was only a
library demo and couldn’t be used as a standalone binary, we transform
it into an example that can be executed using `cargo run --example
parse`.  This also fixes the compiler warnings about unused methods.